refactor: 画像ファイルのリネームと各コンポーネントのimportパス更新 (#105)
- imgs/gallery/sap-activities.jpg → sup-activities.jpg にリネーム
- imgs/gallery/sap-activities2.jpg → sup-activities2.jpg にリネーム
- imgs/ の旧ファイル群（bath-room, life-jacket, living, mah-jongg, sap, table-tennis, wet-suit）を削除
- imgs/ に新ファイル群（facility-*, first-floor-*, rental-*）を追加
- GallerySection.tsx / FeaturesSection.tsx / FacilitiesSection.tsx / PricingSection.tsx: importパスを新ファイル名に更新
- docs/ASSETS.md / docs/IA.md: ファイル名変更に伴う記述を更新
